pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agents-temp"
version = "0.0.1"
description = "A virtual environment for building the agent project"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain>=1.0.3",
    "langchain-openai>=1.0.1",
    "langchain_community>=0.4.1",
    "psycopg[binary,pool]",
    "langgraph",
    "langgraph-checkpoint-postgres",
    "pyyaml>=6.0",
    "python-dotenv>=1.2.1",
    "pydantic==2.12.3",
    "google-api-python-client==2.187.0",
    "beautifulsoup4==4.14.3",
    "qdrant-client==1.16.1",
    "sentence-transformers==5.1.2",
    "pyarrow==22.0.0",
    "pandas==2.3.3",
    "feedparser==6.0.12",
    "tqdm==4.67.1",
    "torch==2.9.1", # To support GPU acceleration, need to install with CUDA support
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.1.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "isort>=5.10.0",
]